+++ core/compat/psp/main.c
@@ -27,9 +27,8 @@ u32 psp_fw_version = 0;
 static ARKConfig _ark_conf;
 ARKConfig* ark_config = &_ark_conf;
 
-extern int is_launcher_mode;
-extern int use_mscache;
-extern void settingsHandler(char* path);
+extern void (*prevPluginHandler)(const char* path, int modid);
+extern void pluginHandler(const char* path, int modid);
 extern void PSPOnModuleStart(SceModule2 * mod);
 
 // Flush Instruction and Data Cache
@@ -62,6 +61,8 @@ int module_start(SceSize args, void * argp)
     processArkConfig(ark_config);
     // Register Module Start Handler
     previous = sctrlHENSetStartModuleHandler(PSPOnModuleStart);
+    // Register plugin handler
+    prevPluginHandler = registerPluginHandler(&pluginHandler);
     // Return Success
     return 0;
 }